% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{code_insee}
\alias{code_insee}
\title{code INSEE}
\format{
A vector
}
\source{
\url{https://www.insee.fr/fr/information/2115000}
}
\usage{
code_insee
}
\description{
A dataset containing insee code of commune
}
\keyword{datasets}
