% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cog_2022}
\alias{cog_2022}
\title{COG 2022}
\format{
A vector
}
\source{
\url{https://www.insee.fr/fr/information/2115000}
}
\usage{
cog_2022
}
\description{
A dataset containing insee code and wording of commune as of January 1, 2022. COG mean
Code Officiel Géographique
}
\keyword{datasets}
