% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cog_2023}
\alias{cog_2023}
\title{COG 2023}
\format{
\subsection{\code{cog_2023}}{

A data frame with 34990 rows and 2 columns:
\describe{
\item{COM}{insee code}
\item{LIBELLE}{Name of commune}
}
}
}
\source{
\url{https://www.insee.fr/fr/information/6800675}
}
\usage{
cog_2023
}
\description{
A dataset containing insee code and wording of commune as of January 1, 2023. COG mean
Code Officiel Géographique
}
\keyword{datasets}
