% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_2024}
\alias{com_2024}
\title{COG 2024}
\format{
\subsection{\code{com_2024}}{

A data frame with 34980 rows and 2 columns:
\describe{
\item{COM}{insee code}
\item{LIBELLE}{Name of commune}
}
}
}
\source{
\url{https://www.insee.fr/fr/information/7766585}
}
\usage{
com_2024
}
\description{
A dataset containing insee code and wording of commune as of January 1, 2024. COG mean
Code Officiel Géographique
}
\keyword{datasets}
