% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apicarto_cadastre.R
\name{get_apicarto_cadastre}
\alias{get_apicarto_cadastre}
\alias{get_apicarto_cadastre.sf}
\alias{get_apicarto_cadastre.sfc}
\alias{get_apicarto_cadastre.character}
\title{Apicarto Cadastre}
\usage{
get_apicarto_cadastre(x, section = NULL, numero = NULL)

\method{get_apicarto_cadastre}{sf}(x, section = NULL, numero = NULL)

\method{get_apicarto_cadastre}{sfc}(x, section = NULL, numero = NULL)

\method{get_apicarto_cadastre}{character}(x, section = NULL, numero = NULL)
}
\arguments{
\item{x}{It can be a shape or a code :
\itemize{
\item Shape Then, all the cadastral parcels contained in it are downloaded. It should be an
object of class \code{sf} or \code{sfc}
\item Code insee : filter the response on the INSEE code entered (must be a \code{character})
}}

\item{section}{A \code{character} to filter the response on the cadastral section code entered (on 2 characters)}

\item{numero}{A \code{character} to filter the answers on the entered parcel number (on 4 characters)}
}
\value{
\code{apicarto_get_cadastre}return an object of class \code{sf}
}
\description{
Apicarto Cadastre
}
\examples{
\dontrun{
library(sf)
library(tmap)

#' # shape from the best town in France
shape <- st_polygon(list(matrix(c(-4.373937, 47.79859,
                                 -4.375615, 47.79738,
                                 -4.375147, 47.79683,
                                 -4.373898, 47.79790,
                                 -4.373937, 47.79859),
                                 ncol = 2, byrow = TRUE)))
shape <- st_sfc(shape, crs = st_crs(4326))

res <- apicarto_get_cadastre(shape)
res2 <- apicarto_get_cadatre("29158")

tm_shape(res)+
   tm_borders()
tm_shape(shape)+
   tm_borders(col = "red")

tm_shape(res2)+
   tm_borders()

}
}
