% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apicarto_gpu.R
\name{get_apicarto_plu}
\alias{get_apicarto_plu}
\title{Apicarto module Geoportail de l'urbanisme}
\usage{
get_apicarto_plu (x,
                  ressource = "zone-urba",
                  partition = NULL)
}
\arguments{
\item{x}{An object of clas \code{sf} or \code{sfc}. If NULL, \code{partition} must be filled by partition of PLU.}

\item{ressource}{A character from this list : "document", "zone-urba", "secteur-cc", "prescription-surf",
"prescription-lin", "prescription-pct", "info-surf", "info-lin", "info-pct". See detail for more info.}

\item{partition}{A character corresponding to PLU partition (can be retrieve using
\code{get_apicarto_plu(x, "document", partition = NULL)}). If partition, is not known, it can be set to
\code{"auto"}. In this case all PLU from the commune is downloaded. If it's set to NULL, only PLU
features intersecting \code{x} are downloaded.}
}
\value{
A object of class \code{sf}
}
\description{
Apicarto module Geoportail de l'urbanisme
}
\details{
\itemize{
\item \verb{"document'} :
\item \code{"zone-urba"} :
\item \code{"secteur-cc"} :
\item \code{"prescription-surf"} :
\item \code{"prescription-lin"} :
\item \code{"prescription-pct"} :
\item \code{"info-surf"} :
\item \code{"info-lin"} :
\item \code{"info-pct"} :
}
}
\examples{
\dontrun{
library(tmap)
library(sf)
point <- st_sfc(st_point(c(-0.4950188466302029, 45.428039987269926)), crs = 4326)

# If you know the partition
partition <- "DU_17345"
poly <- get_apicarto_plu(x = NULL, ressource = "zone-urba", partition = partition)
qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)

# If you don't know partition
poly <- get_apicarto_plu(x = point, ressource = "zone-urba", partition = "auto")
qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)

# If you wanna find partition
document <- get_apicarto_plu(point, ressource = "document", partition = NULL)
partition <- unique(document$partition)

# If you want only intersection between you're shape and PLU
poly <- get_apicarto_plu(point, ressource = "zone-urba", partition = NULL)
qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)

# Get all prescription : /!\ prescription is different than zone-urba

partition <- "DU_17345"
ressources <- c("prescription-surf", "prescription-lin", "prescription-pct")
res <- list()
for (i in 1:3){
    res_temp <- get_apicarto_plu(x, ressources[i], partition)
    res[[i]] <- res_temp
}


}


}
