% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apikeys.R
\name{get_apikeys}
\alias{get_apikeys}
\title{List of all API keys from IGN}
\usage{
get_apikeys()
}
\description{
All apikeys are manually extract from this table provided by IGN :
https://geoservices.ign.fr/documentation/services/tableau_ressources
}
\examples{
get_apikeys()[1]
}
