% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apicarto_gpu.R
\name{get_gpu_layers}
\alias{get_gpu_layers}
\title{Available GPU layers}
\usage{
get_gpu_layers(type = NULL)
}
\arguments{
\item{type}{\code{character} One of \code{"commune"}, \code{"du"}, \code{"prescription"},
\code{"acte-sup"}, \code{"assiette"}, \code{"generateur"}.
If \code{NULL}, all layers are retuned. \code{NULL} by default}
}
\value{
list
}
\description{
Helpers that return available GPU layers and their type.
}
\details{
\code{"du"}: "Document d'urbanisme"
\code{"sup"}: "Servitude d'utilité publique"
}
\examples{
# All layers
names(get_gpu_layers())

# All sup layers
names(get_gpu_layers("generateur"))

# All sup and du layers
names(get_gpu_layers(c("generateur", "prescription")))

}
