% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_news.R
\name{get_last_news}
\alias{get_last_news}
\title{Print latest news from geoservice}
\usage{
get_last_news()
}
\value{
message
}
\description{
Print latest news from geoservice
}
\examples{
\dontrun{
get_last_news()
}
}
