% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_info.R
\name{get_location_info}
\alias{get_location_info}
\title{Retrieve additional information for wms layer}
\usage{
get_location_info(x,
                  apikey = "ortho",
                  layer = "ORTHOIMAGERY.ORTHOPHOTOS",
                  read_sf = FALSE,
                  version = "1.3.0")
}
\arguments{
\item{x}{Object of class \code{sf} or \code{sfc}. Only single point are supported for now.
Needs to be located in France.}

\item{apikey}{\code{character}; API key from \code{get_apikeys()} or directly
from \href{https://geoservices.ign.fr/services-web-experts}{IGN website}.}

\item{layer}{\code{character}; layer name from
\code{get_layers_metadata(apikey, "wms")} or directly from
\href{https://geoservices.ign.fr/services-web-experts}{IGN website}.}

\item{read_sf}{\code{logical}; if \code{TRUE} an \code{sf} object is returned but
response times may be higher.}

\item{version}{\code{character}; version of the service used. See details
for more info.}
}
\value{
\code{character} or \code{sf} containing additional information about the layer
}
\description{
For some wms layer more information can be found with GetFeatureInfo request.
This function first check if info are available. If not, available layers
are returned.
}
\examples{
\dontrun{
library(sf)

# from single point
x <- st_centroid(read_sf(system.file("extdata/penmarch.shp", package = "happign")))
location_info <- get_location_info(x, "ortho", "ORTHOIMAGERY.ORTHOPHOTOS")

# from multiple point


}
}
