% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wfs_layer.R
\name{get_wfs_layer}
\alias{get_wfs_layer}
\title{Download WFS layer}
\usage{
get_wfs_layer(
  shape,
  apikey = "cartovecto",
  layer_name = "BDCARTO_BDD_WLD_WGS84G:troncon_route"
)
}
\arguments{
\item{shape}{Object of class sf. Needs to be located in France}

\item{apikey}{API key from get_apikey() (or directly from the website
https://geoservices.ign.fr/services-web-experts)}

\item{layer_name}{Name of the layer from get_layers_metadata(apikey, "wfs")
(or directly from the website :
https://geoservices.ign.fr/services-web-experts-addAPIkey)}
}
\description{
Directly download a layer from the French National Institute of
Geographic and Forestry. To do that, it need a location giving by
a shapfile, an apikey thanks to get_apikey and the name of layer
thanks to get_layer_metadata.
}
