% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wms_layer.R
\name{get_wms_raster}
\alias{get_wms_raster}
\title{Download WMS raster layer}
\usage{
get_wms_raster(
  shape,
  apikey = "altimetrie",
  layer_name = "ELEVATION.ELEVATIONGRIDCOVERAGE",
  width_height = 500
)
}
\arguments{
\item{shape}{Object of class sf. Needs to be located in France.}

\item{apikey}{API key from get_apikey() (or directly from the website
https://geoservices.ign.fr/services-web-experts)}

\item{layer_name}{Name of the layer from get_layers_metadata(apikey, "wms")
(or directly from the website :
https://geoservices.ign.fr/services-web-experts-addAPIkey)}

\item{width_height}{dim of the raster in pixel}
}
\value{
Spat
}
\description{
Directly download a raster layer from the French National Institute
of Geographic and Forestry. To do that, it need a location giving by
a shapefile, an apikey thanks to get_apikeys() and the name of layer
thanks to get_layer_metadata().
}
