% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happign-package.R
\docType{package}
\name{happign-package}
\alias{happign}
\alias{happign-package}
\title{happign: R Interface to 'IGN' Web Services}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Interface to easily access the National Institute of Geographic and Forestry Information open-source data for any area of interest in France via WFS (shapefile) and WMS (raster) web services <https://geoservices.ign.fr/services-web-experts>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paul-carteron}
  \item \url{https://paul-carteron.github.io/happign/}
  \item Report bugs at \url{https://github.com/paul-carteron/happign/issues}
}

}
\author{
\strong{Maintainer}: Paul Carteron \email{carteronpaul@gmail.com} (\href{https://orcid.org/0000-0002-6942-6662}{ORCID})

}
\keyword{internal}
