% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_2025}
\alias{reg_2025}
\title{French Regions Table (2025)}
\format{
A data frame with one row per region and the following columns:
\describe{
\item{REG}{(int) Region code (2 characters)}
\item{CHEFLIEU_REG}{(chr) Commune code of the regional capital (5 characters)}
\item{TNCC_REG}{(int) Name type indicator (1 character)}
\item{NCC_REG}{(chr) Official name in uppercase (200 characters)}
\item{NCCENR_REG}{(chr) Official name with proper typography (200 characters)}
\item{LIBELLE_REG}{(chr) Official name with article and proper typography (200 characters)}
}
}
\source{
\url{https://www.insee.fr/fr/statistiques/fichier/8377162/cog_ensemble_2025_csv.zip}
}
\usage{
reg_2025
}
\description{
Data for French regions from the INSEE file "Régions".
}
\keyword{datasets}
