% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shp_to_geojson}
\alias{shp_to_geojson}
\title{Convert sf or sfc to geojson format}
\usage{
shp_to_geojson(shape)
}
\arguments{
\item{shape}{A shape of class sf or sfc. Could be a POLYGON, POINT or
LINESTRING.}
}
\value{
Return a geojson string
}
\description{
Convert sf or sfc to geojson format
}
