% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{spatial_cql}
\alias{spatial_cql}
\title{Build a spatial ECQL filter}
\usage{
spatial_cql(x, layer, predicate)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc}}

\item{layer}{\code{character} giving the WFS layer name.}

\item{predicate}{A spatial predicate object created by predicate helpers.}
}
\value{
A character string containing a spatial ECQL filter.
}
\description{
Converts a spatial predicate and a reference geometry into an ECQL
expression suitable for use in a WFS \code{GetFeature} request.
}
\details{
This function is an internal helper used by \code{\link[=get_wfs]{get_wfs()}} to translate
spatial predicate objects (see \link{spatial_predicates}) into ECQL syntax understood by the WFS server.

Users should not call this function directly.
}
\keyword{internal}
