% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_info.R
\name{are_queryable}
\alias{are_queryable}
\title{are_queryable}
\usage{
are_queryable(apikey)
}
\arguments{
\item{apikey}{API key from \code{get_apikeys()} or directly
from the \href{https://geoservices.ign.fr/services-web-experts}{IGN website}}
}
\value{
\code{character} containing the name of the queryable layers
}
\description{
Check if a wms layer is queryable with GetFeatureInfo.
}
\seealso{
\code{\link[=get_location_info]{get_location_info()}}
}
