% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apicarto_cadastre.R
\name{get_apicarto_cadastre}
\alias{get_apicarto_cadastre}
\title{Apicarto Cadastre}
\usage{
get_apicarto_cadastre(x,
                      type = "parcelle",
                      source = "PCI",
                      section = list(NULL),
                      numero = list(NULL),
                      code_arr = list(NULL),
                      code_abs = list(NULL),
                      code_com = list(NULL))
}
\arguments{
\item{x}{It can be a shape, insee codes or departement codes :
\itemize{
\item Shape : must be an object of class \code{sf} or \code{sfc}.
\item Code insee : must be a 5 letters \code{character}
\item Code departement : must be a 2  or 3 letters \code{character}
}}

\item{type}{A \code{character} from "parcelle", "commune", "feuille", "division"}

\item{source}{Can be "BDP" for BD Parcellaire or "PCI" for Parcellaire express.
See detail for more info.}

\item{section}{A \code{character} corresponding to cadastral section}

\item{numero}{A \code{character} corresponding to cadastral number
on the entered parcel number (on 4 characters)}

\item{code_arr}{A \code{character} corresponding to district code for Paris, Lyon, Marseille}

\item{code_abs}{A \code{character} corresponding to the code of absorbed commune.
This prefix is useful to differentiate between communes that have merged}

\item{code_com}{A \code{character} corresponding to the commune code. Only use with
\code{type = "division"} or \code{type = "feuille"}}
}
\value{
Object of class \code{sf}
}
\description{
Implementation of the cadastre module from the
\href{https://apicarto.ign.fr/api/doc/cadastre}{IGN's apicarto}
}
\details{
\code{source}: BD Parcellaire is a discontinued product. Its use is no longer
recommended because it is no longer updated. The use of PCI Express is
strongly recommended and will become mandatory. More information on the
comparison of this two products can be found
\href{https://geoservices.ign.fr/sites/default/files/2021-07/Comparatif_PEPCI_BDPARCELLAIRE.pdf}{here}
}
\examples{
\dontrun{
library(sf)
library(tmap)

}
}
