% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apicarto_gpu.R
\name{get_apicarto_plu}
\alias{get_apicarto_plu}
\title{Apicarto module Géoportail de l'urbanisme}
\usage{
TO DO
}
\arguments{
\item{x}{An object of clas \code{sf} or \code{sfc}. If NULL, \code{partition} must be filled by partition of PLU.}

\item{ressource}{A character from this list : "document", "zone-urba", "secteur-cc", "prescription-surf",
"prescription-lin", "prescription-pct", "info-surf", "info-lin", "info-pct". See detail for more info.}

\item{partition}{A character corresponding to PLU partition (can be retrieve using
\code{get_apicarto_plu(x, "document", partition = NULL)}). If partition, is not known, it can be set to
\code{"auto"}. In this case all PLU from the commune is downloaded. If it's set to NULL, only PLU
features intersecting \code{x} are downloaded.}
}
\value{
A object of class \code{sf}
}
\description{
TO DO DESCRIPTION
}
\details{
\itemize{
\item \verb{"document'} :
\item \code{"zone-urba"} :
\item \code{"secteur-cc"} :
\item \code{"prescription-surf"} :
\item \code{"prescription-lin"} :
\item \code{"prescription-pct"} :
\item \code{"info-surf"} :
\item \code{"info-lin"} :
\item \code{"info-pct"} :
}
}
\examples{
\dontrun{
library(tmap)
library(sf)
point <- st_sfc(st_point(c(-0.4950188466302029, 45.428039987269926)), crs = 4326)
}

}
