% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apikeys.R
\name{get_apikeys}
\alias{get_apikeys}
\title{List of all API keys from IGN}
\usage{
get_apikeys()
}
\value{
character
}
\description{
All API keys are manually extract from this
\href{https://geoservices.ign.fr/documentation/services/tableau_ressources}{table}
provided by IGN.
}
\examples{
\dontrun{
# One API key
get_apikeys()[1]

# All API keys
get_apikeys()

}

}
