% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_news.R
\name{get_last_news}
\alias{get_last_news}
\title{Print latest news from geoservice website}
\usage{
get_last_news()
}
\value{
message or error
}
\description{
This function is a wrapper around the RSS feed of the
geoservice site to get the latest information.
}
\examples{
\dontrun{
get_last_news()
}

}
