% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_news.R
\name{get_last_news}
\alias{get_last_news}
\title{Print latest news from geoservice}
\usage{
get_last_news()
}
\value{
message
}
\description{
This function connects directly to the RSS feed of the
geoservice site to get the latest information. This allows
to understand why some resources are sometimes not available.
}
\examples{
\dontrun{
get_last_news()
}

}
