% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_metadata.R
\name{get_layers_metadata}
\alias{get_layers_metadata}
\alias{get_layers_metadata.character}
\alias{get_layers_metadata.wfs}
\alias{get_layers_metadata.wms}
\title{Metadata for one couple of apikey and data_type}
\usage{
get_layers_metadata(apikey, data_type)

\method{get_layers_metadata}{character}(apikey, data_type)

\method{get_layers_metadata}{wfs}(apikey, data_type)

\method{get_layers_metadata}{wms}(apikey, data_type)
}
\arguments{
\item{apikey}{Apikey from get_apikeys()}

\item{data_type}{Should be "wfs" or "wms}
}
\value{
data.frame with all useful metadata
}
\description{
Metadata are retrieved using the IGN APIs. The execution time can
be long depending on the size of the metadata associated with
the apikey and the overload of the IGN servers.
}
\examples{
\dontrun{
get_layers_metadata("administratif", "wms")
}
}
