% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wfs_attributes.R
\name{get_wfs_attributes}
\alias{get_wfs_attributes}
\title{get_wfs_attributes}
\usage{
get_wfs_attributes(layer = NULL)
}
\arguments{
\item{layer}{\code{character}; name of the WFS layer. Must correspond to a
layer available on the IGN WFS service (see \code{\link[=get_layers_metadata]{get_layers_metadata()}}).}
}
\value{
\code{character}vector with layer attributes
}
\description{
Helper to write ecql filter. Retrieve all attributes from a layer.
}
\examples{
\dontrun{

get_wfs_attributes("LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune")

}
}
