% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wfs_attributes.R
\name{get_wfs_attributes}
\alias{get_wfs_attributes}
\title{get_wfs_attributes}
\usage{
get_wfs_attributes(apikey = NULL, layer_name = NULL, interactive = FALSE)
}
\arguments{
\item{apikey}{API key from \code{get_apikeys()} or directly
from \href{https://geoservices.ign.fr/services-web-experts}{IGN website}}

\item{layer_name}{Name of the layer from \code{get_layers_metadata(apikey, "wfs")}
or directly from
\href{https://geoservices.ign.fr/services-web-experts}{IGN website}}

\item{interactive}{if set to TRUE, no need to specify \code{apikey} and \code{layer_name}, you'll be ask.}
}
\value{
character vector with layers attributes
}
\description{
get_wfs_attributes
}
\examples{
\dontrun{

get_wfs_attributes("administratif", "LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune")

# Interactive session
get_wfs_attributes(interactive = TRUE)

}
}
